+++ packages/clerk-js/src/ui/hooks/useCheckout.ts
@@ -1,5 +1,9 @@
 import { useClerk, useOrganization } from '@clerk/shared/react';
-import type { __experimental_CheckoutProps, __experimental_CommerceCheckoutResource } from '@clerk/types';
+import type {
+  __experimental_CheckoutProps,
+  __experimental_CommerceCheckoutResource,
+  ClerkAPIError,
+} from '@clerk/types';
 import { useCallback, useEffect, useState } from 'react';
 
 import { useFetch } from './useFetch';
@@ -14,6 +18,8 @@ export const useCheckout = (props: __experimental_CheckoutProps) => {
     data: initialCheckout,
     isLoading,
     invalidate,
+    revalidate,
+    error,
   } = useFetch(
     __experimental_commerce?.__experimental_billing.startCheckout,
     {
@@ -25,20 +31,22 @@ export const useCheckout = (props: __experimental_CheckoutProps) => {
     'commerce-checkout',
   );
 
+  const updateCheckout = useCallback((newCheckout: __experimental_CommerceCheckoutResource) => {
+    setCurrentCheckout(newCheckout);
+  }, []);
+
   useEffect(() => {
     if (initialCheckout && !currentCheckout) {
       setCurrentCheckout(initialCheckout);
     }
   }, [initialCheckout, currentCheckout]);
 
-  const updateCheckout = useCallback((newCheckout: __experimental_CommerceCheckoutResource) => {
-    setCurrentCheckout(newCheckout);
-  }, []);
-
   return {
     checkout: currentCheckout || initialCheckout,
     updateCheckout,
     isLoading,
     invalidate,
+    revalidate,
+    isMissingPayerEmail: error?.errors.some((e: ClerkAPIError) => e.code === 'missing_payer_email'),
   };
 };